Commit to MyProfile user story
Commit to MyProfile user story, bonus story for sprint 3. As a user, I want to be able to view all my basic information, as well as all the item that I am selling in the case that I am a vendor. Furthermore, I want to be able to edit said basic information and have them saved and presented.